# src/main/kotlin/com/quodigital/recruit/codetest/bookstore/api/generated/controller/AuthorApi.kt
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
*/
package com.quodigital.recruit.codetest.bookstore.api.generated.controller

import com.quodigital.recruit.codetest.bookstore.api.generated.model.Author
import com.quodigital.recruit.codetest.bookstore.api.generated.model.AuthorRequest
import com.quodigital.recruit.codetest.bookstore.api.generated.model.Error
import io.swagger.v3.oas.annotations.*
import io.swagger.v3.oas.annotations.enums.*
import io.swagger.v3.oas.annotations.media.*
import io.swagger.v3.oas.annotations.responses.*
import io.swagger.v3.oas.annotations.security.*
import org.springframework.http.HttpStatus
import org.springframework.http.MediaType
import org.springframework.http.ResponseEntity

import org.springframework.web.bind.annotation.*
import org.springframework.web.context.request.NativeWebRequest
import org.springframework.beans.factory.annotation.Autowired


import kotlin.collections.List
import kotlin.collections.Map

@RequestMapping("\${api.base-path:}")
interface AuthorApi {

    @Operation(
        tags = ["author",],
        summary = "著者を削除",
        operationId = "authorAuthorIdDelete",
        description = """指定された著者IDの著者を削除します。書籍に紐づいている場合、削除に失敗します。""",
        responses = [
            ApiResponse(responseCode = "204", description = "Success operation"),
            ApiResponse(responseCode = "409", description = "Conflict", content = [Content(schema = Schema(implementation = Error::class))]),
            ApiResponse(responseCode = "200", description = "An unexpected exception occurred", content = [Content(schema = Schema(implementation = Error::class))])
        ]
    )
    @RequestMapping(
            method = [RequestMethod.DELETE],
            value = ["/author/{authorId}"],
            produces = ["application/json"]
    )
    fun authorAuthorIdDelete(@Parameter(description = "著者IDを指定します", required = true) @PathVariable("authorId") authorId: kotlin.Int): ResponseEntity<Unit> {
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
    }

    @Operation(
        tags = ["author",],
        summary = "著者名の変更",
        operationId = "authorAuthorIdPut",
        description = """指定された著者IDの著者名を変更します。変更後の著者名がすでに登録済みである場合はエラーとなります。""",
        responses = [
            ApiResponse(responseCode = "201", description = "Success operation"),
            ApiResponse(responseCode = "400", description = "Bad request", content = [Content(schema = Schema(implementation = Error::class))]),
            ApiResponse(responseCode = "404", description = "Not found", content = [Content(schema = Schema(implementation = Error::class))]),
            ApiResponse(responseCode = "200", description = "An unexpected exception occurred", content = [Content(schema = Schema(implementation = Error::class))])
        ]
    )
    @RequestMapping(
            method = [RequestMethod.PUT],
            value = ["/author/{authorId}"],
            produces = ["application/json"],
            consumes = ["application/json"]
    )
    fun authorAuthorIdPut(@Parameter(description = "著者IDを指定します", required = true) @PathVariable("authorId") authorId: kotlin.Int,@Parameter(description = "", required = true) @RequestBody authorRequest: AuthorRequest): ResponseEntity<Unit> {
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
    }

    @Operation(
        tags = ["author",],
        summary = "著者の検索",
        operationId = "authorAuthorNameGet",
        description = """指定された著者名で著者を検索します。部分一致で検索します。""",
        responses = [
            ApiResponse(responseCode = "200", description = "Success operation", content = [Content(array = ArraySchema(schema = Schema(implementation = Author::class)))]),
            ApiResponse(responseCode = "400", description = "Bad request", content = [Content(schema = Schema(implementation = Error::class))]),
            ApiResponse(responseCode = "404", description = "Not found", content = [Content(schema = Schema(implementation = Error::class))]),
            ApiResponse(responseCode = "200", description = "An unexpected exception occurred", content = [Content(schema = Schema(implementation = Error::class))])
        ]
    )
    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/author/{authorName}"],
            produces = ["application/json"]
    )
    fun authorAuthorNameGet(@Parameter(description = "著者名を指定します", required = true) @PathVariable("authorName") authorName: kotlin.String): ResponseEntity<List<Author>> {
        return ResponseEntity(HttpStatus.NOT_IMPLEMENTED)
    }
}